+++ src/nationalarchives/components/cookie-banner/cookie-banner.stories.js
@@ -6,12 +6,7 @@ import Cookies from "../../lib/cookies.mjs";
 const argTypes = {
   serviceName: { control: "text" },
   cookiesUrl: { control: "text" },
-  policies: { control: "text" },
-  policiesKey: { control: "text" },
   preferencesSetKey: { control: "text" },
-  cookiesDomain: { control: "text" },
-  cookiesPath: { control: "text" },
-  allowInsecure: { control: "boolean" },
   style: {
     control: "inline-radio",
     options: ["none", "contrast", "accent", "tint"],
@@ -34,11 +29,7 @@ export default {
 const Template = ({
   serviceName,
   cookiesUrl,
-  policiesKey,
   preferencesSetKey,
-  cookiesDomain,
-  cookiesPath,
-  allowInsecure,
   style,
   classes,
   attributes,
@@ -47,11 +38,7 @@ const Template = ({
     params: {
       serviceName,
       cookiesUrl,
-      policiesKey,
       preferencesSetKey,
-      cookiesDomain,
-      cookiesPath,
-      allowInsecure,
       style,
       classes,
       attributes,
@@ -70,12 +57,11 @@ export const Accept = Template.bind({});
 Accept.args = {
   serviceName: "My service",
   cookiesUrl: "#",
-  allowInsecure: true,
   classes: "tna-cookie-banner--demo",
   disableMockAnalytics: true,
 };
 Accept.play = async ({ canvasElement }) => {
-  const cookies = new Cookies({ newInstance: true });
+  const cookies = new Cookies({ secure: false, noInit: true });
   await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
   await expect(cookies.isPolicyAccepted("usage")).toEqual(false);
   await expect(cookies.isPolicyAccepted("settings")).toEqual(false);
@@ -105,7 +91,19 @@ Accept.play = async ({ canvasElement }) => {
 
   // await expect(closeButton).not.toBeVisible();
 
-  await cookies.deleteAll();
+  document.cookie.replace(/(?<=^|;).+?(?==|;|$)/g, (name) =>
+    location.hostname
+      .split(".")
+      .reverse()
+      .reduce(
+        (domain) => (
+          (domain = domain.replace(/^\.?[^.]+/, "")),
+          (document.cookie = `${name}=;max-age=0;path=/;domain=${domain}`),
+          domain
+        ),
+        location.hostname,
+      ),
+  );
 };
 
 export const Reject = Template.bind({});
@@ -116,7 +114,7 @@ Reject.args = {
   disableMockAnalytics: true,
 };
 Reject.play = async ({ canvasElement }) => {
-  const cookies = new Cookies({ newInstance: true });
+  const cookies = new Cookies({ secure: false, noInit: true });
   await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
   await expect(cookies.isPolicyAccepted("usage")).toEqual(false);
   await expect(cookies.isPolicyAccepted("settings")).toEqual(false);
@@ -143,34 +141,43 @@ Reject.play = async ({ canvasElement }) => {
   await expect(acceptButton).not.toBeVisible();
   await expect(rejectButton).not.toBeVisible();
 
-  await cookies.deleteAll();
+  document.cookie.replace(/(?<=^|;).+?(?==|;|$)/g, (name) =>
+    location.hostname
+      .split(".")
+      .reverse()
+      .reduce(
+        (domain) => (
+          (domain = domain.replace(/^\.?[^.]+/, "")),
+          (document.cookie = `${name}=;max-age=0;path=/;domain=${domain}`),
+          domain
+        ),
+        location.hostname,
+      ),
+  );
 };
 
 export const Existing = Template.bind({});
 Existing.args = {
   serviceName: "My service",
   cookiesUrl: "#",
-  allowInsecure: true,
   classes: "tna-cookie-banner--demo",
   disableMockAnalytics: true,
 };
 Existing.decorators = [
   (Story) => {
-    const cookies = new Cookies({ newInstance: true });
-    cookies.set(
-      "cookies_policy",
-      JSON.stringify({
-        usage: true,
-        settings: true,
-        marketing: false,
-        essential: false,
-      }),
-    );
+    const cookies = new Cookies({ secure: false, noInit: true });
+    cookies.init();
+    cookies.acceptPolicy("settings");
+    cookies.acceptPolicy("usage");
     return Story();
   },
 ];
 Existing.play = async ({ canvasElement }) => {
-  const cookies = new Cookies();
+  const cookies = new Cookies({
+    newInstance: true,
+    secure: false,
+    noInit: true,
+  });
   await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
   await expect(cookies.isPolicyAccepted("usage")).toEqual(true);
   await expect(cookies.isPolicyAccepted("settings")).toEqual(true);
@@ -184,20 +191,31 @@ Existing.play = async ({ canvasElement }) => {
   await expect(acceptButton).not.toBeVisible();
   await expect(rejectButton).not.toBeVisible();
 
-  await cookies.deleteAll();
+  document.cookie.replace(/(?<=^|;).+?(?==|;|$)/g, (name) =>
+    location.hostname
+      .split(".")
+      .reverse()
+      .reduce(
+        (domain) => (
+          (domain = domain.replace(/^\.?[^.]+/, "")),
+          (document.cookie = `${name}=;max-age=0;path=/;domain=${domain}`),
+          domain
+        ),
+        location.hostname,
+      ),
+  );
 };
 
 export const Partial = Template.bind({});
 Partial.args = {
   serviceName: "My service",
   cookiesUrl: "#",
-  allowInsecure: true,
   classes: "tna-cookie-banner--demo",
   disableMockAnalytics: true,
 };
 Partial.decorators = [
   (Story) => {
-    const cookies = new Cookies({ newInstance: true });
+    const cookies = new Cookies({ secure: false, noInit: true });
     cookies.set(
       "cookies_policy",
       JSON.stringify({
@@ -209,7 +227,11 @@ Partial.decorators = [
   },
 ];
 Partial.play = async ({ canvasElement }) => {
-  const cookies = new Cookies();
+  const cookies = new Cookies({
+    newInstance: true,
+    secure: false,
+    noInit: true,
+  });
   await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
   await expect(cookies.isPolicyAccepted("usage")).toEqual(true);
   await expect(cookies.isPolicyAccepted("settings")).toEqual(false);
@@ -223,26 +245,42 @@ Partial.play = async ({ canvasElement }) => {
   await expect(acceptButton).toBeVisible();
   await expect(rejectButton).toBeVisible();
 
-  await cookies.deleteAll();
+  document.cookie.replace(/(?<=^|;).+?(?==|;|$)/g, (name) =>
+    location.hostname
+      .split(".")
+      .reverse()
+      .reduce(
+        (domain) => (
+          (domain = domain.replace(/^\.?[^.]+/, "")),
+          (document.cookie = `${name}=;max-age=0;path=/;domain=${domain}`),
+          domain
+        ),
+        location.hostname,
+      ),
+  );
 };
 
 export const Malformed = Template.bind({});
 Malformed.args = {
   serviceName: "My service",
   cookiesUrl: "#",
-  allowInsecure: true,
   classes: "tna-cookie-banner--demo",
   disableMockAnalytics: true,
 };
 Malformed.decorators = [
   (Story) => {
-    const cookies = new Cookies({ newInstance: true });
+    document.documentElement.setAttribute("data-tna-cookies-insecure", "true");
+    const cookies = new Cookies({ secure: false, noInit: true });
     cookies.set("cookies_policy", "foobar");
     return Story();
   },
 ];
 Malformed.play = async ({ canvasElement }) => {
-  const cookies = new Cookies();
+  const cookies = new Cookies({
+    newInstance: true,
+    secure: false,
+    noInit: true,
+  });
   await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
   await expect(cookies.isPolicyAccepted("usage")).toEqual(false);
   await expect(cookies.isPolicyAccepted("settings")).toEqual(false);
@@ -256,37 +294,65 @@ Malformed.play = async ({ canvasElement }) => {
   await expect(acceptButton).toBeVisible();
   await expect(rejectButton).toBeVisible();
 
-  await cookies.deleteAll();
+  document.cookie.replace(/(?<=^|;).+?(?==|;|$)/g, (name) =>
+    location.hostname
+      .split(".")
+      .reverse()
+      .reduce(
+        (domain) => (
+          (domain = domain.replace(/^\.?[^.]+/, "")),
+          (document.cookie = `${name}=;max-age=0;path=/;domain=${domain}`),
+          domain
+        ),
+        location.hostname,
+      ),
+  );
 };
 
-//Hidden w. no pref saved
+export const UnexpectedHidden = Template.bind({});
+UnexpectedHidden.args = {
+  serviceName: "My service",
+  cookiesUrl: "#",
+  classes: "tna-cookie-banner--demo",
+  disableMockAnalytics: true,
+};
+UnexpectedHidden.decorators = [
+  (Story) => {
+    document.documentElement.setAttribute("data-tna-cookies-insecure", "true");
+    document.cookie = "cookie_preferences_set=true";
+    return Story();
+  },
+];
+UnexpectedHidden.play = async ({ canvasElement }) => {
+  const cookies = new Cookies({
+    newInstance: true,
+    secure: false,
+    noInit: true,
+  });
+  await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
+  await expect(cookies.isPolicyAccepted("usage")).toEqual(false);
+  await expect(cookies.isPolicyAccepted("settings")).toEqual(false);
+  await expect(cookies.isPolicyAccepted("marketing")).toEqual(false);
+  await expect(cookies.exists("cookie_preferences_set")).toEqual(true);
+  await expect(cookies.get("cookie_preferences_set")).toEqual("false");
 
-// export const EventHandling = Template.bind({});
-// EventHandling.args = {
-//   serviceName: "My service",
-//   cookiesUrl: "#",
-//   cookiesPath: "/tna-frontend/",
-//   policies: "custom",
-//   classes: "tna-cookie-banner--demo",
-// };
-// EventHandling.play = async ({ args, canvasElement }) => {
-//   deleteAllCookies();
+  const canvas = within(canvasElement);
+  const acceptButton = canvas.getByText("Accept cookies");
+  const rejectButton = canvas.getByText("Reject cookies");
+  await expect(acceptButton).toBeVisible();
+  await expect(rejectButton).toBeVisible();
 
-//   const cookies = new Cookies();
-
-//   const onChangePolicy = jest.fn(data => {
-//     console.log(data)
-//   })
-//   cookies.on("changePolicy", onChangePolicy)
-
-//   const canvas = within(canvasElement);
-//   const acceptButton = canvas.getByText("Accept cookies");
-//   await userEvent.click(acceptButton);
-
-//   await expect(onChangePolicy.mock).toHaveBeenCalledTimes(1);
-//   await expect(onChangePolicy.mock.calls).toHaveLength(1);
-//   await expect(onChangePolicy.mock.results[0].value).toHaveProperty("custom");
-//   await expect(onChangePolicy.mock.results[0].value.custom).toEqual(true);
-
-//   deleteAllCookies();
-// };
+  document.cookie.replace(/(?<=^|;).+?(?==|;|$)/g, (name) =>
+    location.hostname
+      .split(".")
+      .reverse()
+      .reduce(
+        (domain) => (
+          (domain = domain.replace(/^\.?[^.]+/, "")),
+          (document.cookie = `${name}=;max-age=0;path=/;domain=${domain}`),
+          domain
+        ),
+        location.hostname,
+      ),
+  );
+};